Worker login Implemented with JWT and tests

+++ DATABASE/src/main/java/com/SFAE/SFAE/Controller/WorkerController.java
@@ -10,10 +10,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
+import com.SFAE.SFAE.DTO.LoginRequest;
 import com.SFAE.SFAE.DTO.WorkerDTO;
 import com.SFAE.SFAE.ENDPOINTS.WorkerEp;
 import com.SFAE.SFAE.ENTITY.Worker;
 import com.SFAE.SFAE.INTERFACE.WorkerInterface;
+import com.SFAE.SFAE.Security.JWT;
 import com.SFAE.SFAE.Service.Authentication;
 
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
@@ -31,11 +33,14 @@ public class WorkerController implements WorkerEp {
     @Autowired
     private WorkerInterface dao;
 
-    @Autowired
-    private Authentication login;
+   @Autowired 
+   private JWT jwt;
 
     @Override
     public ResponseEntity<Worker> createWorker(@RequestBody WorkerDTO worker) {
+        if(worker==null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         try{
             Worker builded=dao.createWorker(worker);
             
@@ -49,6 +54,9 @@ public class WorkerController implements WorkerEp {
 
     @Override
     public ResponseEntity<?> deleteWorkerById(long id) {
+        if(id<0){
+            return ResponseEntity.badRequest().body("id can be <0 or is null");
+        }
         try{
             boolean result=dao.deleteWorkerById(id);
             if(result){
@@ -69,7 +77,7 @@ public class WorkerController implements WorkerEp {
             return ResponseEntity.status(HttpStatus.OK).body(worker);
        }
        catch(Exception e){
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
        }
       
      
@@ -78,7 +86,7 @@ public class WorkerController implements WorkerEp {
     @Override
     public ResponseEntity<?> findWorkersbyID(long id) {
         if(id<0){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            return ResponseEntity.badRequest().body("id can be <0 or is null");
         }
             Worker found=dao.findWorkersbyID(id);
 
@@ -89,7 +97,7 @@ public class WorkerController implements WorkerEp {
     @Override
     public ResponseEntity<?> findWorkerByName(String name) {
         if(name.length()<0){
-            throw new IllegalArgumentException("Name not found for "+name);
+            return ResponseEntity.badRequest().body("Name is null");
         }
           Worker found=dao.findWorkerbyName(name);
           return ResponseEntity.status(HttpStatus.OK).body(found);
@@ -97,22 +105,32 @@ public class WorkerController implements WorkerEp {
 
     @Override
     public ResponseEntity<Worker> updateWorker(@RequestBody WorkerDTO jsonData) {
+        if(jsonData ==null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         try{
             dao.updateWorker(jsonData);
             
         }
         catch(Exception e){
             System.out.println(e);
-            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
         return ResponseEntity.status(HttpStatus.ACCEPTED).build();
     }
 
     @Override
-    public ResponseEntity<?> loginWorker(String email, String password) {
+    public ResponseEntity<?> loginWorker(@RequestBody LoginRequest login) {
+        if(login.getEmail()== null||login.getPassword() == null){
+            ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
         try{
-            login.loginWorker(email, password);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).build();
+            String token=jwt.loginWorkerJWT(login.getEmail(),login.getPassword());
+            if(token==null){
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
+            return ResponseEntity.ok().body(token);
         }
        catch(Exception e) {
         return ResponseEntity.status(HttpStatus.NON_AUTHORITATIVE_INFORMATION).build();
